Setup Webpacker and ReactRails
This PR sets up Webpacker with ReactRails, to allow a more flexible frontend dev experience. This uses css-modules, react, the usual. You can see I also converter the first section of the home page to react, so everybody can see how it should be done. Notice the prerender part, this allows the page to be used without React and so the server can render raw HTML. The objective is to have a component for each page (atm you can see it's just HomeHeader, in the future, we should have a HomePage component rendering each section).